add handlerfn, router and structs file
