+++ src/structs/articleStructs.js
