+++ src/structs/productStructs.js
